requirements.txt
PyQt6>=6.4.0
uvicorn>=0.20.0
fastapi>=0.88.0
setuptools>=60.2.0
websockets>=10.4
websocket-client>=1.4.2